0. Armazem de items

+++ Generics_Exercises/armazem/Comida.java
@@ -1,11 +1,11 @@
 public class Comida {
 
     private String nome;
-    private int calorias;
+    private double calorias;
     private double preco;
 
 
-    public Comida(String nome, int calorias, double preco) {
+    public Comida(String nome, double calorias, double preco) {
         this.calorias = calorias;
         this.nome = nome;
         this.preco = preco;
@@ -13,6 +13,6 @@ public class Comida {
 
     @Override
     public String toString() {
-        return "[%s] %d R$ %f".formatted(nome, calorias, preco);
+        return "[%s] %f R$ %f".formatted(nome, calorias, preco);
     }
 }
